# openapi.yaml
openapi: 3.1.0
info:
    title: Prometheus API
    description: Prometheus is an Open-Source monitoring system with a dimensional data model, flexible query language, efficient time series database and modern alerting approach.
    contact:
        name: Prometheus Community
        url: https://prometheus.io/community/
    version: 3.11.2
servers:
    - url: http://localhost:9090/api/v1
paths:
...
    /targets:
        get:
            tags:
                - targets
            summary: Get targets
            operationId: get-targets
            parameters:
                - name: scrapePool
                  in: query
                  description: Filter targets by scrape pool name.
                  required: false
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: prometheus
                - name: state
                  in: query
                  description: 'Filter by state: active, dropped, or any.'
                  required: false
                  explode: false
                  schema:
                    type: string
                  examples:
                    example:
                        value: active
            responses:
                "200":
                    description: Target discovery information retrieved successfully.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/TargetsOutputBody'
                            examples:
                                targetsList:
                                    summary: Active and dropped targets
                                    value:
                                        data:
                                            activeTargets:
                                                - discoveredLabels:
                                                    __address__: demo.prometheus.io:9093
                                                    __meta_filepath: /etc/prometheus/file_sd/alertmanager.yml
                                                    __metrics_path__: /metrics
                                                    __scheme__: http
                                                    env: demo
                                                    job: alertmanager
                                                  globalUrl: http://demo.prometheus.io:9093/metrics
                                                  health: up
                                                  labels:
                                                    env: demo
                                                    instance: demo.prometheus.io:9093
                                                    job: alertmanager
                                                  lastError: ""
                                                  lastScrape: "2026-01-02T13:36:40.200Z"
                                                  lastScrapeDuration: 0.006576866
                                                  scrapeInterval: 15s
                                                  scrapePool: alertmanager
                                                  scrapeTimeout: 10s
                                                  scrapeUrl: http://demo.prometheus.io:9093/metrics
                                            droppedTargetCounts:
                                                alertmanager: 0
                                                blackbox: 0
                                                caddy: 0
                                                cadvisor: 0
                                                grafana: 0
                                                node: 0
                                                prometheus: 0
                                                random: 0
                                            droppedTargets: []
                                        status: success
                default:
                    description: Error retrieving targets.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                            examples:
                                tsdbNotReady:
                                    summary: TSDB not ready
                                    value:
                                        error: TSDB not ready
                                        errorType: internal
                                        status: error
...
components:
    schemas:
        Error:
            type: object
            properties:
                status:
                    type: string
                    enum:
                        - success
                        - error
                    description: Response status.
                    example: success
                errorType:
                    type: string
                    description: Type of error that occurred.
                    example: bad_data
                error:
                    type: string
                    description: Human-readable error message.
                    example: invalid parameter
            required:
                - status
                - errorType
                - error
            additionalProperties: false
            description: Error response.
        Labels:
            type: object
            additionalProperties: true
            description: Label set represented as a key-value map.
...
        Target:
            type: object
            properties:
                discoveredLabels:
                    $ref: '#/components/schemas/Labels'
                labels:
                    $ref: '#/components/schemas/Labels'
                scrapePool:
                    type: string
                    description: Name of the scrape pool.
                scrapeUrl:
                    type: string
                    description: URL of the target.
                globalUrl:
                    type: string
                    description: Global URL of the target.
                lastError:
                    type: string
                    description: Last error message from scraping.
                lastScrape:
                    type: string
                    format: date-time
                    description: Timestamp of the last scrape.
                lastScrapeDuration:
                    type: number
                    format: double
                    description: Duration of the last scrape in seconds.
                health:
                    type: string
                    description: Health status of the target (up, down, or unknown).
                scrapeInterval:
                    type: string
                    description: Scrape interval for this target.
                scrapeTimeout:
                    type: string
                    description: Scrape timeout for this target.
            required:
                - discoveredLabels
                - labels
                - scrapePool
                - scrapeUrl
                - globalUrl
                - lastError
                - lastScrape
                - lastScrapeDuration
                - health
                - scrapeInterval
                - scrapeTimeout
            additionalProperties: false
            description: Scrape target information.
        DroppedTarget:
            type: object
            properties:
                discoveredLabels:
                    $ref: '#/components/schemas/Labels'
                scrapePool:
                    type: string
                    description: Name of the scrape pool.
            required:
                - discoveredLabels
                - scrapePool
            additionalProperties: false
            description: Dropped target information.
        TargetDiscovery:
            type: object
            properties:
                activeTargets:
                    type: array
                    items:
                        $ref: '#/components/schemas/Target'
                droppedTargets:
                    type: array
                    items:
                        $ref: '#/components/schemas/DroppedTarget'
                droppedTargetCounts:
                    type: object
                    additionalProperties:
                        type: integer
                        format: int64
            required:
                - activeTargets
                - droppedTargets
                - droppedTargetCounts
            additionalProperties: false
            description: Target discovery information including active and dropped targets.
        TargetsOutputBody:
            type: object
            properties:
                status:
                    type: string
                    enum:
                        - success
                        - error
                    description: Response status.
                    example: success
                data:
                    $ref: '#/components/schemas/TargetDiscovery'
                warnings:
                    type: array
                    items:
                        type: string
                    description: Only set if there were warnings while executing the request. There will still be data in the data field.
                infos:
                    type: array
                    items:
                        type: string
                    description: Only set if there were info-level annotations while executing the request.
            required:
                - status
                - data
            additionalProperties: false
            description: Response body for targets endpoint.
...